Handle log2, log10 more robustly (#355)
* Gemini: log2, log10 handling improvement

* Gemini: Inline LOG264/LOG1064 in logbase to preserve original function ordering

* Gemini: Add explanatory comments for logarithm routines and macros

* Improve comments

* delint

* Fine-tune NEWS

* Gemini: Add test coverage for general base vector and scalar cases in log.integer64

+++ src/integer64.h
@@ -225,7 +225,7 @@ else {                                                   \
     else { \
         if (e1 < 0) \
             naflag = TRUE; \
-        ret = (double) sqrt((long double)e1); \
+        ret = (double) sqrtl((long double)e1); \
     }
 
 #define LOG64(e1, ret, naflag) \
@@ -237,11 +237,12 @@ else {                                                   \
             naflag = TRUE; \
     }
 
+// NB: cast to double _after_ dividing in 'long double' for max precision.
 #define LOGVECT64(e1, e2, ret, naflag) \
     if (e1 == NA_INTEGER64) \
         ret = NA_REAL; \
     else { \
-        ret = (double) logl((long double)e1) / log(e2); \
+        ret = (double) (logl((long double)e1) / logl((long double)e2)); \
         if (isnan(ret)) \
             naflag = TRUE; \
     }
@@ -250,7 +251,7 @@ else {                                                   \
     if (e1 == NA_INTEGER64) \
         ret = NA_REAL; \
     else { \
-        ret = (double) logl((long double)e1) / e2; \
+        ret = (double) (logl((long double)e1) / e2); \
         if (isnan(ret)) \
             naflag = TRUE; \
     }
